+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 + [x] Animation: fixed using String.includes for IE11. Change to String.indexOF.
 + [x] Animation: add operator `/` to `_getRelativeValue()`.
 + [x] Events: fix `fire`. Now main is constructor `CustonEvent`, for old - `createEvent`
-+ [x] Manipulation: add method `wrap`, `wrapAll`. The method puts elements inside the wrapper. 
++ [x] Manipulation: add method `wrap`, `wrapAll`, `wrapInner`. The method puts elements inside the wrapper and return `wrapper(s)`. 
++ [x] Init: change `throw Error` to `console.warn` when selector is `#` or `.`
 
 ### 1.0.6
 + [x] Init: fix minor bug for creating elements in context